requirements.txt
blinker==1.9.0
certifi==2026.4.22
charset-normalizer==3.4.7
click==8.1.8
contourpy==1.3.0
cycler==0.12.1
Flask==3.1.3
fonttools==4.60.2
geopandas==1.0.1
idna==3.13
importlib_metadata==8.7.1
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.7
MarkupSafe==3.0.3
matplotlib==3.9.4
numpy==2.0.2
packaging==26.2
pandas==2.3.3
pillow==11.3.0
pyogrio==0.11.1
pyparsing==3.3.2
pyproj==3.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2026.1.post1
requests==2.32.5
scipy==1.13.1
shapely==2.0.7
six==1.17.0
tzdata==2026.2
urllib3==2.6.3
Werkzeug==3.1.8
zipp==3.23.1
